Refactoring users table definition.
Refactoring Users table definition and some typo fixes.

	modified:   src/html/index.html
	modified:   src/python/CertWebServer.py
	modified:   src/python/apache_utils.py
